Added location service that AddCemeteryFragment.kt binds to to get cemetery location for the user.
Added sendCemsToNework which allows user to add a new cemetery to the server database.

If network request fails cemetery is added to database work manager handles sending cemetery to network later.

CemeteryRefreshWorker.kt gets most recent cemetery epoch time added and compares with local databases time added. If they are out of sync if makes a network request to sync. Still need to test work manager

Setup sending a new grave to the network

When adding a grave it is inserted into local db then sent to network. When MyCemeteriesFragment.kt is shown it pull the local graves that user has added.